requirements.txt
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0
requests>=2.31
pybaseball>=2.2
MLB-StatsAPI>=1.7
yfinance>=0.2.40
matplotlib>=3.7
plotly>=5.18
jupyter>=1.0
pytest>=7.4
